Revert "Merge pull request #106 from AlexOxytocin/codex/void-core-fixes"
This reverts commit 911d70185ab442bce69f3eb9980842314b299477, reversing
changes made to 519bb712bcfb98ef95e2d6a223bb2e66878d2ab7.

+++ src/community_bot/application/moderation.py
@@ -19,7 +19,6 @@ from community_bot.domain.moderation import (
 )
 
 if TYPE_CHECKING:
-    from decimal import Decimal
     from types import TracebackType
 
     from community_bot.application.identity import ActorContext
@@ -46,7 +45,7 @@ class ModerationCaseDetail:
     case: ModerationCase
     task_title: str
     task_origin: str
-    credit_reward_per_performer: Decimal
+    credit_reward_per_performer: int
     assignment_status: str
     result_summary: str | None
     dispute_reason: str
